Fix kicad-data.cabal so that it builds (#1)
I attempted to build kicad-data, and I ran into a couple of problems:

First, building kicad-data from Hackage does not work, because Data/Kicad/Util.hs is missing from the tarball. In order for it to be included, it needs to be listed in kicad-data.cabal.

Second, the lower bound for parsec is too loose, because endOfLine did not appear in parsec until version 3.1.6.

This pull request fixes both of these issues. Thanks!